+++ src/main/java/dev/muon/irons_apothic/affix/SpellLevelAffix.java
@@ -130,8 +130,7 @@ public class SpellLevelAffix extends SchoolFilteredAffix {
     }
 
     private static Component formatSchool(SchoolType school) {
-        String key = "school." + school.getId().getNamespace() + "." + school.getId().getPath();
-        return Component.translatable(key).withStyle(school.getDisplayName().getStyle());
+        return school.getDisplayName();
     }
 
     @Override